Add game function and prompt for user input
Added a game function which begins a 5 round game of rock paper
scissors. Game function prompts the user for input at the beginning
of each round, runs computerPlay, and then compares the two selections
to determine a winner of each round. The game function keeps track of
each player's score and determines a winner (or draw) at the end of 5
rounds.

Additionally the code was cleaned up slightly with some comments
describing the use of each function.

Minor changes were made to the playRound function to print the results
message directly to console and then change the return value to an
easier-to-deal-with string for the purposes of keeping track of score
in the game function.